Track the producer of every published number, and correct a wall that was two walls
The site publishes figures whose producers are absent from the history: the
block-summary error bar, the overflow ceiling, the gate-kill counts, the fold
census windows and the estimated-operator margins are all emitted by files a
reader cannot reach. This change tracks those producers, retracts a constant
that was wrong on the page, inverts three tests whose green required a defect
to remain, and makes the figure contract fail loudly where it previously
deferred in silence. The operator's frozen three-switch head is unchanged.

Provenance

  63 previously untracked paths enter the history, including the producers of
  every number the site carries: ceqjepa/fold_census.py for the three stability
  windows, tests/curvature/test_block_summary_binds.py for the 2.085e-15
  E-EXACT bar, tests/arm_smprime/test_bf16_ceiling.py for the overflow walls,
  tests/arm_smprime/test_gate_kill_mask.py for the nan counts, and
  tests/cameron/test_estimated_operator_margin.py for the sigma_total grading.

The overflow wall

  The page stated that bfloat16 and float32 share an overflow wall at
  88.72283935546875. They do not. They share an 8-bit exponent, but bfloat16
  carries fewer mantissa bits, so its largest finite value is smaller and it
  overflows first:

    float32  max 3.4028234663852886e+38   log 88.72283905206835
    bfloat16 max 3.3895313892515355e+38   log 88.71892521235186

  At x = 88.7215 float32 reads 3.3982583016316987e+38 and bfloat16 reads inf;
  the two disagree across [88.71892521235186, 88.72283905206835]. The published
  figure was not even the correct float32 value, differing in the eighth
  decimal. The surviving claim is stronger than the retracted one: bfloat16
  does not widen the ceiling, it narrows it. Both constants are now computed
  from torch.finfo at test time rather than carried as literals, and
  tests/arm_smprime/test_bf16_ceiling.py gains an ordering assertion,
  bf16_wall < f32_wall, that the retracted claim would have failed.

Tests that enforced defects

  Three tests asserted behaviour that the gate-kill repair removed, so their
  green depended on the defect remaining and each fired on whoever fixed it.
  tests/arm_smprime/test_bf16_ceiling.py asserted nan on the live diagonal and
  now asserts inf_only > 0; tests/curvature/test_bucket_numerics.py asserted
  mod[2, 0] != mod[2, 0] on a dead entry and now asserts it is exactly 0.0,
  confirmed red when the two-line dead-entry guard is removed in a scratch copy
  (mod[2] reads [nan, 0.0, 2.3245822930325816e-50] without it). A tree-wide
  census over every isnan, isinf and defect-named test found no further
  instances, and distinguished thirteen that resemble the pattern but guard
  correct edge cases: an all-masked row genuinely has no value, and asserting
  nan there is right.

Figure contract

  The markup and the drawing code both supplied viewBox, slider bounds and
  control names, and whichever ran second won in silence. Three rounds of
  figure defects trace to that one mechanism. shell() and range() now assert on
  disagreement rather than defer, which surfaced six pre-existing mismatches on
  the unmodified page at once. Measured on mkdocs build output in headless
  Chromium: 0 of 351 fold-slider positions disagree with their own label,
  against 184 before; documentElement scrollWidth equals clientWidth at 375x812.
  tests/site/test_figures_do_not_lie.py runs rather than skips on this machine
  and enforces one rule, that no label may claim a count the drawing does not
  show.

Validation

  python -m pytest --collect-only -q reads 3,438 tests collected, 8 errors, all
  eight rooted in torchvision 0.20.1 standing against torch 2.14.0 and reaching
  transformers through ceq/hf/modeling_ceq.py. requirements.txt now names those
  eight files and ships a re-derive command that reproduces them; the command it
  carried previously returned zero files and contradicted its own comment.
  Across every directory this change touches, python -m pytest reads 171 passed
  in 71.46s on WIN-16QAL06O9GB, python 3.11.9, torch 2.14.0+cpu.
  python -m mkdocs build --strict exits 0 in 19.12s and node --check passes on
  docs/assets/resolvent.js.

What the measurements killed

  Eleven claims died with the number that killed them, each recorded rather
  than removed. The U(1) phase is gauge-trivial on every bed here, removable by
  a tree gauge to 9.602627e-15 at 2.61x the cost of a real signed gate. A
  softmax stack leaks 0.000074 across a genuine two-closed-class wall against
  its own R2 seed spread of 0.0043, 58x below noise. Effective rank prefers a
  frozen-random encoder over a K-cluster representation by 22.08x at D=64, so
  both geometric collapse legs are inverted. A hitting-time bed has no live cell
  at any budget: its unscoreable gate sits flat at 0.89 across four decades
  because it counts the sealed share of the draw, and its refusal head is beaten
  7.3x on recall by a one-line non-causal shortcut that reads no structure.

Limits

  The structural claim that the attention family cannot represent a closed class
  is retracted: sparsemax opens n closed classes by direct construction, and the
  Lean theorems bind exp-based scores only. Cross-row consistency survives at
  k=1 with a hand-set boundary, 1500/1500 against sparsemax's 1352/1500, but a
  learned gate is not bitwise zero under powers, 224 of 12,000 values nonzero in
  4 of 5 seeds, and the leak that grows with k is generic mixing rather than the
  boundary corridor proposed for it. C9's reference triple is definitional
  rather than predicted, since the pinned bed's diagonal Q makes r_gamma a
  one-line identity. The prediction leg of the north star remains unmet: every
  exact-read win was handed the true operator, and the one honest estimated arm
  has mle_gap 0.0 exactly, making it transition counting. The 160-goal
  depth-race sweep is running and unfinished, and its numbers compare only
  within themselves because pool 80 and pool 160 share 38 of 100 test goals.

+++ ceq/arm_pl.py
@@ -312,7 +312,8 @@ def brute_force_path_sums(a: torch.Tensor) -> torch.Tensor:
     """
     m = a.detach().cpu().numpy()
     n = m.shape[0]
-    r = np.eye(n)
+    r = np.eye(n, dtype=m.dtype)  # was np.eye(n) (float64): silently
+    # truncated the imaginary part of `total` on assignment for complex `A`.
     for i in range(n):
         for j in range(i):
             total = 0.0
